LayoutTests/media/media-source/media-source-paint-to-canvas-expected.txt

EVENT(sourceopen)
EVENT(update)
EVENT(canplay)
RUN(video.currentTime += 1.001 / 24)
EVENT(seeked)
EXPECTED (canvas.getContext("2d").getImageData(250, 130, 1, 1).data[0] != '0') OK
RUN(video.currentTime += 1.001 / 24)
EVENT(seeked)
EXPECTED (canvas.getContext("2d").getImageData(250, 130, 1, 1).data[0] != '0') OK
RUN(video.currentTime += 1.001 / 24)
EVENT(seeked)
EXPECTED (canvas.getContext("2d").getImageData(250, 130, 1, 1).data[0] != '0') OK
END OF TEST
